refactor(server): delete dead backward-compat shims
- Delete tunnel.js (re-export shim for tunnel/cloudflare.js)
- Delete ws-file-ops.js (re-export shim for ws-file-ops/index.js)
- Remove dead re-exports from ws-message-handlers.js (constants and
  functions that are canonical in handler-utils.js)
- Delete tunnel.test.js (tested the deleted shim)
- Update 10 test files to import from canonical module paths
- Update architecture docs to reflect removed files

Closes #2514
