+++ docs/architecture/reference.md
@@ -53,12 +53,11 @@ For essential dev workflow, see [CLAUDE.md](/CLAUDE.md).
 | ToolResult | `src/tool-result.js` | Tool result processing and formatting |
 | TunnelCheck | `src/tunnel-check.js` | Tunnel health verification (DNS propagation) |
 | TunnelEvents | `src/tunnel-events.js` | Tunnel event wiring helpers |
-| TunnelManager | `src/tunnel.js` | Backward-compat shim re-exporting CloudflareTunnelAdapter |
 | TunnelRegistry | `src/tunnel/registry.js` | Tunnel adapter registry (`registerTunnel`/`getTunnel`/`parseTunnelArg`) |
 | WebTaskManager | `src/web-task-manager.js` | Claude Code Web cloud task management |
 | WsBroadcaster | `src/ws-broadcaster.js` | Message broadcast to session and global scopes |
 | WsClientManager | `src/ws-client-manager.js` | Client connection lifecycle management |
-| WsFileOps | `src/ws-file-ops.js` | File browsing/reading WS message handlers |
+| WsFileOps | `src/ws-file-ops/index.js` | File browsing/reading WS message handlers |
 | WsForwarding | `src/ws-forwarding.js` | Session event → WS broadcast wiring |
 | WsMessageHandlers | `src/ws-message-handlers.js` | WS message handler dispatch |
 | WsPermissions | `src/ws-permissions.js` | Permission request/response WS message handlers |
