--- packages/server/src/tunnel.js
@@ -1,5 +1,0 @@
-/**
- * Backward-compatibility re-export.
- * New code should import from './tunnel/index.js' instead.
- */
-export { CloudflareTunnelAdapter as TunnelManager } from './tunnel/cloudflare.js'
