--- packages/server/src/ws-file-ops.js
@@ -1,2 +1,0 @@
-// Re-export from modular implementation for backward compatibility
-export { createFileOps } from './ws-file-ops/index.js'
